fix for paths starting with ../
Fix #33
Re: IndigoUnited/node-cross-spawn-async#19

Previously, paths starting with `./` were shortened to remove the `./`
portion.  However, the regular expression was too loose, and only
checked that it started with `.`, but not that it had a slash after a
single `.`.

There is still a weird inconsistency with `which(1)` in this area,
though:

    $ node bin/which ././bin/which
    ./bin/which

    $ which ././bin/which
    ././bin/which

I don't *think* this would cause any problems, but technically speaking,
it's incorrect, so should be patched later.  Most likely, we're cleaning
something that shouldn't be getting cleaned, but I have not investigated
it yet.