Add PATHEXT tests
The tests `find when executable > with pathExt` were setting
`process.env.PATHEXT`, as if to test to ensure that it would read from
it correctly. But in both tests, the option `pathExt` was also set.
This option is checked before `PATHEXT`, so `PATHEXT` was not tested.

The `addTest` function has been given an additional optional parameter
to specify a non-default set of options. New tests have been added to
ensure `PATHEXT` is read correctly.