Fix Node.js v4 build, add Node.js v8
When using Node.js v4, the command `npm install -g npm` will install a
version of npm incompatible with Node.js v4. As a workaround, that step
has been omitted for Node.js v4, and instead the default version of npm
is used.

Node.js v8 as been added as well, as it is the current LTS version of
Node.js.